BaseConnectionHandler: fix crash on destruction
When destroying a BaseConnectionHandlers internal Connection object, it
tries to flush any remaining data.  But by doing so, it attempts to use
methods of derived classes which at this point have already been
destroyed.

The fix is to explicitly destroy the Connection object from a derived
class before running the destructor of BaseConnectionHandler.