Fix Agents publishing messages on closed channels
Agents might not be aware that they are trying to publish messages on a
channel whose underlying connection was closed.  This results in errors
being thrown by `aiormq`.  We improve the situation for management and
data connections in a two-step process:

1. Any time a connection closes, it clears an `asyncio.Event`;  once the
   connection is reestablished it is set again.  Agents that need to
   publish to a channel on these connections then `await` this event
   before actually publishing, thus minimizing the risk of using a
   connection that is closed.
2. It might happen that a connection closes again right after it was
   reestablished.  An Agent then successfully waits for the reconnect-
   event to be set, only to publish on an connection that is now closed
   again.  We cannot prevent that from happening, so we raise exceptions
   accordingly (see below).

We define a base exception `PublishFailedError` from which exceptions
for concrete situations derive:

* If sending values for a metric failed, we raise `MetricSendError`.
* If issuing an RPC request failed, we raise `RpcRequestError`.
* If replying to an RPC request failed, we raise `RpcReplyError`.

The source exceptions (since aiormq 3.1.0: `ChannelInvalidStateError`)
for these errors are attached as a cause.