Allow Sinks to receive metric data after reconnect
Previously, Sinks whose data connection was severed stopped receiving
data after a reconnect.  This was for two reasons.  Firstly, the manager
declares the data queue for a Sink as "auto-delete", i.e. after a Sink
disconnects, it is deleted.  Secondly, the client declared the data as
"robust", so it would be redeclared after a reconnect.

Together, this results in a Sink subscribing to a data queue which does
not receive values anymore, as it is not bound to any producers by the
manager.

To fix this, a Sink needs to resubscribe to the metrics it is interested
via RPC `sink.subscribe`.  This way, the manager is asked to provision a
new data queue, from which we start consuming data for all the requested
metrics.

When implementing this functionality there are some technical
considerations to be made:

* To know whether the data connection was reestablished successfully, we
  override the `DataClient._on_data_connection_reconnect`-callback with
  our own `Sink._on_data_connection_reconnect`-callback.  The task of
  the original callback is to signal when the data connection is ready
  to be used again.  For Sink, we instead use it to first schedule the
  resubscription process and then, after successful completion, trigger
  the original callback from there.
* Previously, the Sink did not track the metrics it was subscribed to.
  In order to be able to reconnect without the user noticing, we need to
  track all metrics that are `Sink.subscribe`d and `Sink.unsubscribe`d.
  This is done in the set `Sink._subscribed_metrics`.
* Because resubscribing is an `async` task (spawned into the event loop
  from the synchronous reconnect-callback) the Sink needs to keeps track
  of it in `self._resubscribe_task`.
* Before a reconnect event triggers the resubscription, any ongoing
  resubscription task is considered to be out-of-date and cancelled.
  Only then a new resubscription task is scheduled.
* Before starting to consume from the new data queue acquired from
  `sink.subscribe`, we need to cancel the old consumer.  This might
  result in a `RuntimeError: <Channel "..." closed>`.  `aiormq` fixes
  this to be a more intuitive error in `3.1.0`, but we need to wait for
  `aio-pika` to incorporate that.