Unify discover responses for C++ and python clients
When receiving to a `discover` broadcast RPC, a client will now respond
with a JSON object in the form of

    interface DiscverResponse {
        alive?: bool;
        currentTime?: Iso8601String;
        startingTime?: Iso8601String;
        uptime?: NanoSeconds;
    }

When receiving such a response, note that all fields are optional.  If
`alive` is missing, assume the client is alive; if `uptime` is missing
but both `currentTime` and `startingTime` are given, calculate `uptime`
as the difference (in number of nanoseconds) between the latter two.

This closes #55 (Agree on discover RPC response structure) and fixes #47
(Discover for python clients fails with datetime error).